Added BMDK Project
Added BMDK (BIRT Mobile Development Kit).
Currently has getters and setters for username, password, hostname, and
volume name.  All are encrypted in a private variable.  Also contains an
ENUM for output types that will be used in the stubbed out method
exportReport() and other methods that will be added later.